Wire Route A into the real DriveVA pipeline and prove it on released weights
The Route A module existed but had never touched the real model.  This adds the
guarded insertion point and fixes the three defects that only a real bf16
30-block pipeline could expose.

Integration
- `model_fn_wan_video` gains a duck-typed `dit._tokenpress_route_a` branch at the
  exact point where x / t_mod / freqs / t are ready and before the dense block
  loop.  When the attribute is absent it is a no-op, so every existing path is
  bit-identical (323 framework tests still pass).  It reuses the real
  `dit.head` and `dit.unpatchify`, so Route A's recovered dense video grid goes
  through the same output path as the dense model.

Defects found only by running it (each would have silently poisoned a GPU run)
1. fp32/bf16 promotion: the gate works in fp32 and `video * mask` promoted the
   whole residual stream to fp32, so every downstream block mismatched the bf16
   weights.  The mask (and the recovery decoder output feeding the head) is now
   cast back to the model dtype.
2. Per-token flow timestep: DriveVA passes a length-N timestep where conditioned
   history tokens carry 0 and future tokens carry sigma.  The scorer required a
   scalar or length-B value and asserted.  Per-token timesteps are now supported
   natively -- which is strictly better, because a token's own noise level is
   exactly the context the plan wants the scorer to condition on.
3. Un-batched positions: `build_driveva_video_positions` returns [N,3] and the
   recovery decoder indexes it with a [B,K,3] index, so the omitted-positions
   path raised "Index tensor must have the same number of dimensions".  The
   synthetic tests all passed explicit positions and never hit it.

Verification
- `scripts/route_a_real_forward_smoke.py`: loads the released pdms90_9 checkpoint
  through the official harness, attaches the 177.5M-parameter Route A module to
  the live DiT (dim=3072, 30 blocks, bf16, 1569-token sequence) and runs a real
  NAVSIM scene.  Result: VALID_RECORD=True, 1.5 s/scene trajectory-only.
  The scorer is randomly initialised so the PDM is 0.0 -- this is a wiring test,
  not a quality result, and the script says so.
- 59 Route A unit tests and the full 323-test suite pass.

+++ videopress_framework/videopress/retraining/route_a.py
@@ -395,8 +395,12 @@ class RouteADynamicSelect(nn.Module):
             )
         else:
             positions = positions.to(device=x.device, dtype=video_hidden.dtype)
-            if positions.ndim == 2:
-                positions = positions.unsqueeze(0).expand(x.shape[0], -1, -1)
+        # Always batch the coordinates.  ``build_driveva_video_positions`` returns
+        # an unbatched [N, 3] table, and the recovery decoder indexes it with a
+        # [B, K, 3] index, so leaving it 2-D raises "Index tensor must have the
+        # same number of dimensions as input tensor" on the real pipeline.
+        if positions.ndim == 2:
+            positions = positions.unsqueeze(0).expand(x.shape[0], -1, -1)
 
         # -------------------------------------------------------- thresholding
         logits = self.scorer(
@@ -443,6 +447,11 @@ class RouteADynamicSelect(nn.Module):
 
         # ----------------------------------------------------- sparse backend
         # Scale the video tokens by the straight-through mask.
+        # The gate works in fp32 for numerical stability, but the backend
+        # blocks run in the model dtype.  Cast the mask back before it touches
+        # the residual stream, otherwise ``video * mask`` silently promotes the
+        # sequence to fp32 and every downstream layer raises a dtype mismatch.
+        ste_mask = ste_mask.to(video_hidden.dtype)
         gated_video = video_hidden * ste_mask.unsqueeze(-1)
         x_gated = torch.cat([gated_video, traj_hidden], dim=1)
 
@@ -524,6 +533,9 @@ class RouteADynamicSelect(nn.Module):
                 query_positions=positions,
                 batch_size=x.shape[0],
             )
+            # Same reason as the mask: the recovery decoder is fp32, the Wan
+            # head runs in the model dtype.
+            dense_video_hidden = dense_video_hidden.to(video_hidden.dtype)
             if head is not None:
                 # The Wan head is position-wise over the *video* grid only: the
                 # production pipeline strips the trajectory tokens before
